+++ projects/core/src/global-message/http-interceptors/handlers/forbidden.handler.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { HttpErrorHandler } from './response.handler';
+import { HttpErrorHandler } from './http-error.handler';
 import { GlobalMessageType } from '../../models/global-message.model';
 import { HttpResponseStatus } from '../../models/response-status.model';
